requirements.txt
Django==2.2.11
djangorestframework
django-cors-headers
mysqlclient
gunicorn
httplib2
daphne
channels
channels_redis
websocket
